backend/ai-service/requirements.txt
# Core dependencies
fastapi>=0.103.1,<0.104.0
uvicorn>=0.23.2,<0.24.0
pydantic>=2.3.0,<2.4.0
pydantic-settings>=2.0.3,<2.1.0
python-dotenv>=1.0.0,<1.1.0
python-jose[cryptography]==3.3.0
passlib>=1.7.4,<1.8.0
bcrypt>=4.0.1,<4.1.0

# Database
sqlalchemy>=2.0.21,<2.1.0
psycopg2-binary>=2.9.7,<2.10.0
alembic>=1.12.0,<1.13.0
asyncpg==0.28.0

# Cache & Message Queue
redis>=5.0.0,<5.1.0
aioredis==2.0.1
celery==5.3.4

# API client
httpx>=0.24.1,<0.25.0
aiohttp==3.8.6

# AI & ML
openai>=0.28.0,<0.29.0
transformers>=4.33.2,<4.34.0
torch>=2.0.1,<2.1.0
numpy>=1.24.3,<1.25.0
scikit-learn>=1.3.0,<1.4.0
pandas>=2.1.0,<2.2.0
nltk>=3.8.1,<3.9.0
spacy>=3.6.1,<3.7.0

# Cryptography & Security
cryptography>=41.0.3,<41.1.0
pyjwt>=2.8.0,<2.9.0
python-multipart>=0.0.6,<0.1.0

# Monitoring & Metrics
prometheus-client>=0.17.1,<0.18.0
sentry-sdk>=1.31.0,<1.32.0

# AWS
boto3==1.28.64
aioboto3==11.3.0

# Testing
pytest>=7.4.2,<7.5.0
pytest-asyncio>=0.21.1,<0.22.0
pytest-cov>=4.1.0,<4.2.0

# Utilities
python-dateutil>=2.8.2,<2.9.0
pytz>=2023.3,<2024.0
email-validator>=2.0.0,<2.1.0
Jinja2==3.1.2
Markdown==3.5
beautifulsoup4==4.12.2
tenacity>=8.2.3,<8.3.0
cachetools>=5.3.1,<5.4.0 